+++ lib/flutterfall.dart
@@ -1,3 +1,4 @@
-library flutterfall;
+library;
 
 export 'src/fall_widget.dart';
+export 'src/fall_controller.dart';